Update pictureNFT.move
There are several errors, bugs, and typos in the code above. One of them is the import of the  Self  module from the  std::string  crate. It should be importing the  String  type instead. The correct import statement should be:

 
use std::string::{String, ToString};
 

Additionally, in the  buy_picture  function, the  ENoPicture  error code is being used incorrectly. It should be replaced with  EPictureExists  since the picture being bought is already listed for sale. Finally, in the  transfer_picture  function, the condition to check if the picture is listed for sale should be removed since it is not relevant to the function's purpose.

I have also found another error in the function  update_picture . The function allows for the update of a picture's details even if the picture is listed for sale, which should not be permitted. To fix this error, you should add an assertion to check if the picture is listed for sale before allowing it to be updated. The revised function should look like this:

 
public fun update_picture(
        gallery: &mut Gallery,
        picture_id: UID,
        new_uri: String,
        new_price: u64,
        ctx: &mut TxContext,
    ) {
        let picture = table::borrow_mut(&mut gallery.pictures, picture_id);
        assert!(picture.owner == tx_context::sender(ctx), ENotOwner);
        assert!(!picture.for_sale, EListedForSale); // Don't allow updates if listed for sale
        picture.uri = new_uri;
        picture.price = new_price;
    }
 

With this assertion, the function now checks if the picture is listed for sale before allowing it to be updated. If the picture is listed for sale, the function will throw an error. This change ensures that the picture's details are not updated while it is still listed for sale.